assets/I8_문서저장서버/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-multipart==0.0.9
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
